app/requirements.txt
fastapi~=0.116.0
uvicorn~=0.35.0
sqlalchemy~=2.0.41
jinja2~=3.1.6
passlib[bcrypt]~=1.7.4
python-jose~=3.5.0
python-dotenv~=1.1.1
psycopg2-binary
pydantic[email]
pydantic-settings~=2.10.1
starlette~=0.46.2
alembic
itsdangerous
python-multipart
